Refactor authy
